+++ NEWS.md
@@ -3,7 +3,9 @@
 * `cloud_check(r_version = "4.3.1")` is the updated default (#361).
 
 * `cloud_check()` gains the ability to check Bioconductor packages via a new
-  `bioc` argument, with default `TRUE` (#362)
+  `bioc` argument, with a default of `FALSE` do to a relatively high likelihood
+  of failed checks since Bioconductor system dependencies are currently not
+  installed in the cloud check service (#362, #369).
 
 * updated pkgdown template and url to https://revdepcheck.r-lib.org.
 
@@ -14,7 +16,7 @@
 
 * `cloud_results()` gains a progress bar so you can see what's happening
   for large revdep runs (#273)
-  
+
 * `cloud_report()` can opt-out of saving failures, and saves the CRAN report
   the same way as `revdep_report()` (#271).